// src/game/types/sprite-keys.type.ts
export const enum TEXTURE_KEYS {
  PLAYER_CAR_LEFT = "PLAYER_CAR_LEFT",
  PLAYER_CAR_CENTER = "PLAYER_CAR_CENTER",
  PLAYER_CAR_RIGHT = "PLAYER_CAR_RIGHT",
  ROAD = "ROAD",
  SKY = "SKY",
  MOUNTAIN_FADE = "MOUNTAIN_FADE",
  MOUNTAIN_LEFT = "MOUNTAIN_LEFT",
  MOUNTAIN_RIGHT = "MOUNTAIN_RIGHT",
  SIDEROAD_LEFT = "SIDEROAD_LEFT",
  SIDEROAD_RIGHT = "SIDEROAD_RIGHT",
}